+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix E2E savings tests failing due to incorrect card click (menu button required)
+
 ## [1.2.1] - 2025-12-25
 
 ### Changed